replication: replicate pastes to all other instances.
With this setup you can run the pastebin behind a load balancer, on more than
one machine, with more than one physical disk.

List all the hosts in the pool in the config (even yourself), and when a paste
is posted, it will be PUT to each of the backends, who will write it to disk.

If no peers are listed, we simply write it to disk as usual, though named
after the sha1 of their content.